requirements.txt
tensorflow
pandas
matplotlib
tqdm
Pillow
scikit-image
opencv-python
nltk
seaborn